+++ src-tauri/src/tools/check_online.rs
@@ -1,7 +1,8 @@
 use crate::Result;
-use crate::chat::{Content, IncomingContent, call_openrouter};
+use crate::chat::Content;
 use crate::error::Error;
-use crate::tools::{Function, Tool};
+use crate::openrouter::{IncomingContent, Openrouter};
+use crate::tools::{Function, Tool, ToolPayload};
 use schemars::{JsonSchema, schema_for};
 use serde::{Deserialize, Serialize};
 use serde_json::{from_str, json, to_value};
@@ -22,7 +23,7 @@ pub struct CheckOnlineResult {
     citations: Vec<String>,
 }
 
-pub async fn check_online(args: CheckOnlineArgs) -> Result<CheckOnlineResult> {
+pub async fn check_online(args: CheckOnlineArgs) -> Result<ToolPayload> {
     let settings = task::spawn_blocking(crate::get_settings_fn).await??;
 
     if settings.api_key.is_empty() {
@@ -46,9 +47,10 @@ pub async fn check_online(args: CheckOnlineArgs) -> Result<CheckOnlineResult> {
     ]
     });
     let schema = to_value(schema_for!(CheckOnlineResult)).unwrap(); // unwrap: all input controlled by code
-    let response = call_openrouter(&[message], SEARCH_MODEL, "", &vec![], Some(schema)).await?;
+    let response = Openrouter::call(&[message], SEARCH_MODEL, "", &vec![], Some(schema)).await?;
     if let IncomingContent::Text(text) = &response.choices[0].message.content {
-        return Ok(from_str(text)?);
+        let result: CheckOnlineResult = from_str(text)?;
+        return ToolPayload::from(result);
     }
     Err(Error::Tool(
         "Failed to get a valid response from the online search tool.".to_string(),
