+++ src-tauri/src/tools/notes.rs
@@ -6,15 +6,16 @@ use std::io::Write;
 use std::path::PathBuf;
 use tokio::task;
 
-use crate::tools::Function;
 use crate::tools::Tool;
+use crate::tools::{Function, ToolPayload};
 
-pub async fn read_notes() -> crate::Result<String> {
+pub async fn read_notes() -> crate::Result<ToolPayload> {
     let notes_path = get_notes_path().await?;
     if !notes_path.exists() {
-        return Ok("No notes found.".to_string());
+        return ToolPayload::from("No notes found.".to_string());
     }
-    fs::read_to_string(notes_path).map_err(|e| Error::Tool(e.to_string()))
+    let result = fs::read_to_string(notes_path).map_err(|e| Error::Tool(e.to_string()))?;
+    ToolPayload::from(result)
 }
 
 #[derive(Deserialize)]
@@ -22,7 +23,7 @@ pub struct AppendNotesArgs {
     pub markdown_content: String,
 }
 
-pub async fn append_notes(args: AppendNotesArgs) -> crate::Result<String> {
+pub async fn append_notes(args: AppendNotesArgs) -> crate::Result<ToolPayload> {
     let notes_path = get_notes_path().await?;
     let timestamp = Local::now().format("%Y-%m-%d %H:%M:%S");
     let note_entry = format!(
@@ -39,7 +40,7 @@ pub async fn append_notes(args: AppendNotesArgs) -> crate::Result<String> {
     file.write_all(note_entry.as_bytes())
         .map_err(|e| Error::Tool(e.to_string()))?;
 
-    Ok("Note appended successfully.".to_string())
+    ToolPayload::from("Note appended successfully.".to_string())
 }
 
 async fn get_notes_path() -> crate::Result<PathBuf> {
