+++ src-tauri/src/tools/roll_dice.rs
@@ -1,4 +1,4 @@
-use crate::tools::{Function, Tool};
+use crate::tools::{Function, Tool, ToolPayload};
 use rand::Rng;
 use serde::Deserialize;
 
@@ -20,7 +20,7 @@ pub fn get_tool() -> Tool {
     }
 }
 
-pub async fn execute(_args: RollDiceArgs) -> u8 {
+pub async fn execute(_args: RollDiceArgs) -> ToolPayload {
     tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-    rand::rng().random_range(1..=6)
+    ToolPayload::from(rand::rng().random_range(1..=6)).unwrap()
 }
